Accounted for gender data errors
There are times when gender data doesn't come through as a simple F or M. If it's anything other than the expected values, R script converts it to NA. But we can't import NA into our system. It has to be either F, M or blank. So we have to leave it blank. This new line makes sure that if the value is NA, its converted to a blank instead.